requirements.txt
Flask
flask-restful
gunicorn
tensorflow >= 2.4.1
pandas
Pillow
Keras
